+++ CHANGELOG.md
@@ -75,6 +75,7 @@ Breaking changes will be documented in this changelog file for each version.
 * github-382: [BUG] Parallel execution of multiple dq checks runs too many checks on Java 17
 * github-384: Improve documentation for using docker-compose
 * github-377: Load override config/env from .flowman-env.yml
+* github-344: Support .flowman-ignore file for Flowman Kernel client
 
 ### Breaking changes